RAT-301: Change example to non-meaningful Java to not find this class in IDE searches in RAT's codebase
